Fixed prefix attack vector
Hey Jix,
This is a gotcha that we can prevent. When Jixbot is a moderator, commands with a leading $0$ can be used to do a number of privileged commands that we probably don't want users to be allowed to do such as purge/timeout/banning other users or exploiting another channel bot.

The goal of this change is to stop commands like this:
!addcommand !<name> $0$ <stuff>
from being added. I've seen commands like this happen before, always by mistake.

I'm assuming that T.response has the whitespace trimmed from the beginning. This probably isn't the optimal way to check a string prefix, but I don't know golang.