Generalize to prevent chosen prefixes
I'm allowing a leading $u$ because I can't think of when a username wouldn't be already escaped enough. If you decide against that you could do `^\$[0-9u]\$` for the regex string similar to how you have for the argument validation below it.

Tested the regex against a few sample strings in the go playground.